refactor: move components, blocks and bases into one component (bricks)
