--- components/polylith/base/__init__.py
@@ -1,4 +1,0 @@
-from polylith.base.create import create_base
-from polylith.base.get import get_bases_data
-
-__all__ = ["create_base", "get_bases_data"]
