--- components/polylith/block/__init__.py
@@ -1,3 +1,0 @@
-from polylith.block.blocks import create_block
-
-__all__ = ["create_block"]
