--- components/polylith/component/__init__.py
@@ -1,5 +1,0 @@
-from polylith.component.create import create_component
-from polylith.component.get import get_components_data
-
-
-__all__ = ["create_component", "get_components_data"]
