+++ components/polylith/poetry_command/create_base.py
@@ -1,7 +1,7 @@
 from cleo.helpers import option
 from poetry.console.commands.command import Command
+from polylith.bricks import base
 from polylith.poetry_command.create import create
-from polylith import base
 
 
 class CreateBaseCommand(Command):
